refactor: update HeaderRegionsDropdown import path and remove unused component feat: add StrategiseResults component for displaying strategy results
